requirements.txt
Pillow
easyocr
torch
requests
